+++ case/case-infra/src/main/java/net/bakaar/sandbox/cas/infra/spring/controler/CaseDTO.java
@@ -1,4 +1,4 @@
-package net.bakaar.sandbox.cas.infra.controler;
+package net.bakaar.sandbox.cas.infra.spring.controler;
 
 import net.bakaar.sandbox.cas.domain.Case;
