+++ case/case-infra/src/main/java/net/bakaar/sandbox/cas/infra/spring/repository/CaseEntity.java
@@ -1,4 +1,4 @@
-package net.bakaar.sandbox.cas.infra.repository.springdata;
+package net.bakaar.sandbox.cas.infra.spring.repository;
 
 import net.bakaar.sandbox.cas.domain.Case;
 import net.bakaar.sandbox.cas.domain.vo.PNummer;
