+++ case/case-infra/src/test/java/net/bakaar/sandbox/cas/infra/spring/CaseInfraConfigurationTest.java
@@ -1,4 +1,4 @@
-package net.bakaar.sandbox.cas.infra;
+package net.bakaar.sandbox.cas.infra.spring;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
