+++ case/case-infra/src/test/java/net/bakaar/sandbox/cas/infra/spring/controler/CaseDTOTest.java
@@ -1,4 +1,4 @@
-package net.bakaar.sandbox.cas.infra.controler;
+package net.bakaar.sandbox.cas.infra.spring.controler;
 
 import org.junit.Test;
