+++ case/case-infra/src/test/java/net/bakaar/sandbox/cas/infra/spring/repository/CaseEntityTest.java
@@ -1,4 +1,4 @@
-package net.bakaar.sandbox.cas.infra.repository.springdata;
+package net.bakaar.sandbox.cas.infra.spring.repository;
 
 import net.bakaar.sandbox.cas.domain.Case;
 import net.bakaar.sandbox.cas.domain.vo.PNummer;
